Split util into type util and epxression util
