+++ src/common/types/builtin.ts
@@ -1,3 +1,4 @@
+import { intInterval, interval, literal } from './type-util';
 import {
     IntIntervalType,
     IntervalType,
@@ -10,7 +11,6 @@ import {
     ValueType,
 } from './types';
 import { union } from './union';
-import { intInterval, interval, literal } from './util';
 
 const fixRoundingError = (n: number): number => {
     if (!Number.isFinite(n) || Number.isInteger(n)) return n; // no fixing necessary
