+++ src/common/types/intersection.ts
@@ -1,5 +1,6 @@
 /* eslint-disable no-continue */
 import { assertNever, sameNumber } from '../util';
+import { groupByUnderlying, isSameStructType } from './type-util';
 import {
     AnyType,
     IntIntervalType,
@@ -15,7 +16,6 @@ import {
     ValueType,
 } from './types';
 import { unionValueTypes } from './union';
-import { groupByUnderlying, isSameStructType } from './util';
 
 const intersectInterval = (
     a: IntervalType | IntIntervalType,
