+++ src/common/types/union.ts
@@ -1,6 +1,7 @@
 /* eslint-disable no-continue */
 import { assertNever, sameNumber } from '../util';
 import { canonicalize } from './canonical';
+import { isSameStructType } from './type-util';
 import {
     AnyType,
     CanonicalTypes,
@@ -19,7 +20,6 @@ import {
     UnionType,
     ValueType,
 } from './types';
-import { isSameStructType } from './util';
 
 type NonEmptyArray<T> = [T, ...T[]];
